feat(archive-node-adapter): refactor events and actions into services
